+++ typings/xlsxwriter/format.py
@@ -1,0 +1,6 @@
+# SPDX-FileCopyrightText: 2024-2025 Stefan Bellon
+#
+# SPDX-License-Identifier: MIT
+
+
+class Format: ...