+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # ChangeLog
 
+## Unreleased
+
+### Changed
+- replaced 'openpyxl' with 'xlsxwriter' to reduce dependencies
+
 ## 0.8.0 (2025-04-19)
 
 ### Added